Fixes some errors.
In case of connection time test, ping the first participant to avoid timeout of the driver session, as collecting data is taking some time.
Skip FileNotFound errors on failures when using remote driver.

+++ src/test/java/org/jitsi/meet/test/base/FailureListener.java
@@ -348,9 +348,13 @@ public class FailureListener
 
             if (type == ParticipantFactory.ParticipantType.chrome)
             {
+                File srcFile = new File(outputLogsParentFolder,
+                    "chrome" + suffix + ".log");
+                // in case of remote driver the file does not exist
+                if (!srcFile.exists())
+                    return;
                 FileUtils.copyFile(
-                    new File(outputLogsParentFolder,
-                        "chrome" + suffix + ".log"),
+                    srcFile,
                     new File(outputLogsParentFolder,
                         fileNamePrefix + suffix + "-chrome" + extension)
                 );
